% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{add_mark_generic}
\alias{add_mark_generic}
\title{Add a visual mark at the current line}
\usage{
add_mark_generic(number)
}
\arguments{
\item{number}{A single digit from 0 to 9 indicating the mark number.}
}
\description{
Adds a numbered mark (0-9) to the current line of the active document,
and inserts a visual comment like \verb{#  MARK N}. If a previous mark with
the same number exists, it will be removed and replaced.
}
\keyword{internal}
