% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{clean_all_marks}
\alias{clean_all_marks}
\title{Remove all visual marks from the current document}
\usage{
clean_all_marks()
}
\description{
Deletes all \verb{# <LABEL> N} comments and clears memory of saved marks.
}
