% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\docType{data}
\name{.mark_env}
\alias{.mark_env}
\title{Internal environment to store mark positions}
\format{
An object of class \code{environment} of length 2.

An object of class \code{environment} of length 2.
}
\usage{
.mark_env

.mark_env
}
\description{
Internal environment to store mark positions

Internal environment to store mark positions and label
}
\keyword{internal}
