% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{goto_mark_generic}
\alias{goto_mark_generic}
\title{Jump to a previously set mark}
\usage{
goto_mark_generic(number)
}
\arguments{
\item{number}{A single digit from 0 to 9.}
}
\description{
Moves the cursor to the line where the specified mark was saved.
If the mark doesn't exist, nothing happens.
}
\keyword{internal}
